+++ src/debugAdapter/activateMplabxDebug.ts
@@ -9,19 +9,41 @@
 
 import * as vscode from 'vscode';
 import { WorkspaceFolder, DebugConfiguration, ProviderResult, CancellationToken } from 'vscode';
-import { MplabxDebugSession } from './mplabxDebug';
 import { FileAccessor } from '../common/FileAccessor';
+import { DebugProtocol } from '@vscode/debugprotocol';
+import { ILaunchRequestArguments } from './mplabxDebug';
+import { MplabxConfigFile } from '../common/configFileHelper';
+import path = require('path');
+import fs = require('fs');
 
-export function activateMplabxDebug(context: vscode.ExtensionContext, factory?: vscode.DebugAdapterDescriptorFactory) {
+
+export function activateMplabxDebug(context: vscode.ExtensionContext, debugType: string, factory: vscode.DebugAdapterDescriptorFactory) {
 
 	// register a configuration provider for 'mplabx' debug type
 	const provider = new MPLABXConfigurationProvider();
-	context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider('mplabx', provider));
+	context.subscriptions.push(vscode.debug.registerDebugConfigurationProvider(debugType, provider));
 
-	if (!factory) {
-		factory = new InlineDebugAdapterFactory();
-	}
-	context.subscriptions.push(vscode.debug.registerDebugAdapterDescriptorFactory('mplabx', factory));
+	context.subscriptions.push(vscode.debug.registerDebugAdapterDescriptorFactory(debugType, factory));
+}
+
+/**
+ * This interface describes the mock-debug specific launch attributes
+ * (which are not part of the Debug Adapter Protocol).
+ * The schema for these attributes lives in the package.json of the mock-debug extension.
+ * The interface should always match this schema.
+ */
+interface ILaunchProjectRequestArguments extends DebugProtocol.LaunchRequestArguments {
+	/** An absolute path to the project to debug. */
+	program: string;
+
+	/** The name of the configuration to debug */
+	configuration?: string;
+
+	/** Automatically stop target after launch. If not specified, target does not stop. */
+	stopOnEntry?: boolean;
+
+	/** Boolean indicating whether to build and launch a debug build or a production build */
+	debug?: boolean;
 }
 
 class MPLABXConfigurationProvider implements vscode.DebugConfigurationProvider {
@@ -32,13 +54,117 @@ class MPLABXConfigurationProvider implements vscode.DebugConfigurationProvider {
 	 */
 	resolveDebugConfiguration(folder: WorkspaceFolder | undefined, config: DebugConfiguration, token?: CancellationToken): ProviderResult<DebugConfiguration> {
 
-		if (!config.program) {
-			return vscode.window.showInformationMessage("Cannot find a program to debug").then(_ => {
-				return undefined;	// abort launch
-			});
+		switch (config.type) {
+			case 'mplabx': default: {
+				let projectConfig = config as unknown as ILaunchProjectRequestArguments;
+				if (folder) {
+					projectConfig.program = projectConfig.program.replace('${workspaceFolder}', folder.uri.path);
+				}
+				
+				let result = this.convertProjectArgs(projectConfig) as unknown as DebugConfiguration;
+				return result;
+			}
+			case 'mdb': {
+
+			}
 		}
 
 		return config;
+	}
+
+	/** A dictionary translating Configuration tool names to MDB tool names */
+	private confToMdBNames = {
+		"PICkit3PlatformTool": "PICKit3",
+		"pk4hybrid": "PICKit4",
+		"pkob4hybrid": "PKOB4",
+		"PK5Tool": "pickit5",
+		"ri4hybrid": "ICE4",
+		"ICD3Tool": "icd3",
+		"ICD4Tool": "icd4",
+		"ICD5Tool": "icd5",
+		"Simulator": "Sim",
+		"jlink": "jlink",
+	};
+
+	private convertProjectArgs(args: ILaunchProjectRequestArguments): ILaunchRequestArguments | undefined{
+
+		try {
+			
+			let project = MplabxConfigFile.readSync(args.program);
+			let confs = project.confs.conf;
+
+			let conf = confs.find(c => c.name === args.configuration);
+
+			if (!conf) {
+				throw Error(`Failure to find a "${args.configuration}" configuration in ${project}`);
+			}
+
+			// Get the tool
+			let tool = conf.toolsSet.platformTool;
+			const device: string = conf.toolsSet.targetDevice;
+
+			// Find the elf
+			let outputFolder = path.join(args.program, 'dist', args.configuration ? args.configuration : 'default', args.debug ? 'debug' : 'production');
+
+			const fileType: string = '.elf';
+
+			if (fs.existsSync(outputFolder)) {
+				let outputFiles = fs.readdirSync(outputFolder, { withFileTypes: true })
+					.filter(item => item.isFile())
+					.filter(item => path.extname(item.name) === fileType);
+
+				if (outputFiles.length > 0) {
+					let outputFile = outputFiles[0].name;
+
+					const programerAllowArray = vscode.workspace.getConfiguration('vslabx').get<string[]>('programerToolAllowList');
+					const programerAllowRegExp: RegExp | undefined = programerAllowArray && programerAllowArray.length > 0 ?
+						RegExp(`(${programerAllowArray?.join('|')})`) : undefined;
+
+					let toolOptions: any = {};
+
+					// Collect all the tool settings
+					if (conf[tool] && programerAllowRegExp) {
+						const allowRegex: RegExp = programerAllowRegExp;
+
+						conf[tool].property.forEach((pair) => {
+							const key: string = pair.key;
+							// Keys with a capital value don't work
+							if (key.toLowerCase() === key) {
+								const value: string = pair.value;
+
+								// Values with '{' in it, needs resolved... idk how to do
+								if (value.length > 0 && !value.match(/(\$\{.+\}|Press\sto|system settings|\.\D)/) &&
+									key.match(allowRegex)) {
+
+									toolOptions[key] = value;
+								}
+							}
+						});
+					}
+
+					// Convert from a project name to an MDB name
+					if (tool in this.confToMdBNames) {
+						tool = this.confToMdBNames[tool];
+					}
+
+					// Add all the properties to the result to make it work like the mdb args
+					args["device"] = device;
+					args["toolType"] = tool;
+					args["filePath"] = path.join(outputFolder, outputFile);
+					args["toolOptions"] = toolOptions;
+					args["stopOnEntry"] = args.stopOnEntry;
+
+					return args as unknown as ILaunchRequestArguments;
+
+				} else {
+					throw new Error(`Failure to find a "${fileType}" file to write to device`);
+				}
+			} else {
+				throw new Error(`Failure to find the output folder: ${outputFolder}`);
+			}
+		} catch (e: any) {
+			vscode.window.showErrorMessage(e.message);
+		}
 	}
 }
 
@@ -66,9 +192,3 @@ function pathToUri(path: string) {
 	}
 }
 
-class InlineDebugAdapterFactory implements vscode.DebugAdapterDescriptorFactory {
-
-	createDebugAdapterDescriptor(_session: vscode.DebugSession): ProviderResult<vscode.DebugAdapterDescriptor> {
-		return new vscode.DebugAdapterInlineImplementation(new MplabxDebugSession(workspaceFileAccessor));
-	}
-}